requirements.txt
Django==1.6
Pillow==2.2.1
PyYAML==3.10
South==0.8.4
argparse==1.2.1
distribute==0.7.3
django-colorful==0.1.3
django-mptt==0.6.0
django-override-settings==1.2
django-permission==0.4.5
icalendar==3.5
mock==1.0.1
python-dateutil==2.2
pytz==2013.8
setuptools-git==1.0
six==1.4.1
wsgiref==0.1.2
